+++ src/loaders.py
@@ -1,14 +1,13 @@
 import os
-from math import sqrt
 import torch
 from attrdict import AttrDict
 from sklearn.externals import joblib
 from torch.utils.data import Dataset, DataLoader
 from torch.utils.data.sampler import Sampler
 from PIL import Image
+from steppy.base import BaseTransformer
 
-from .steps.base import BaseTransformer
-from .utils import meshgrid, box_iou, box_nms, change_box_order
+from .retinanet import DataEncoder
 
 
 class RandomSubsetSampler(Sampler):
@@ -24,7 +23,7 @@ class RandomSubsetSampler(Sampler):
         return self.subset_size
 
 
-class DetectionDataset(Dataset):
+class ImageDetectionDataset(Dataset):
     def __init__(self, X, y, labels, images_dir, encoder, train_mode):
         super().__init__()
         self.X = X
@@ -84,14 +83,14 @@ class DetectionDataset(Dataset):
         return torch.FloatTensor(boxes), torch.LongTensor(labels)
 
     def collate_fn(self, batch):
-        '''Encode targets.
+        """Encode targets.
 
         Args:
           batch: (list) of images, cls_targets, loc_targets.
 
         Returns:
           images, stacked cls_targets, stacked loc_targets.
-        '''
+        """
         imgs = [x[0] for x in batch]
         boxes = [x[1] for x in batch]
         labels = [x[2] for x in batch]
@@ -106,7 +105,7 @@ class DetectionDataset(Dataset):
         return inputs, torch.stack(loc_targets), torch.stack(cls_targets)
 
 
-class ImageSegmentationLoaderBasic(BaseTransformer):
+class ImageDetectionLoader(BaseTransformer):
     def __init__(self, labels, loader_params, dataset_params):
         super().__init__()
         self.labels = labels
@@ -114,7 +113,7 @@ class ImageSegmentationLoaderBasic(BaseTransformer):
         self.dataset_params = AttrDict(dataset_params)
 
         self.encoder = DataEncoder()
-        self.dataset = None
+        self.dataset = ImageDetectionDataset
 
     def transform(self, X, y, X_valid=None, y_valid=None, train_mode=True):
         if train_mode and y is not None:
@@ -137,6 +136,10 @@ class ImageSegmentationLoaderBasic(BaseTransformer):
                                    images_dir=self.dataset_params.images_dir,
                                    encoder=self.encoder,
                                    train_mode=True)
+
+            datagen = DataLoader(dataset, **loader_params,
+                                 sampler=RandomSubsetSampler(len(dataset), self.loader_params.training.subset_size),
+                                 collate_fn=dataset.collate_fn)
         else:
             dataset = self.dataset(X, y,
                                    labels=self.labels,
@@ -144,7 +147,8 @@ class ImageSegmentationLoaderBasic(BaseTransformer):
                                    encoder=self.encoder,
                                    train_mode=False)
 
-        datagen = DataLoader(dataset, **loader_params, sampler=RandomSubsetSampler(len(dataset), 10))
+        datagen = DataLoader(dataset, **loader_params,
+                             collate_fn=dataset.collate_fn)
         steps = len(datagen)
         return datagen, steps
 
@@ -156,123 +160,3 @@ class ImageSegmentationLoaderBasic(BaseTransformer):
     def save(self, filepath):
         params = {'loader_params': self.loader_params}
         joblib.dump(params, filepath)
-
-
-class DataEncoder(object):
-    def __init__(self):
-        self.anchor_areas = [32*32., 64*64., 128*128., 256*256., 512*512.]  # p3 -> p7
-        self.aspect_ratios = [1/2., 1/1., 2/1.]
-        self.scale_ratios = [1., pow(2,1/3.), pow(2,2/3.)]
-        self.anchor_wh = self._get_anchor_hw()
-
-    def _get_anchor_hw(self):
-        '''Compute anchor height and width for each feature map.
-
-        Returns:
-          anchor_hw: (tensor) anchor hw, sized [#fm, #anchors_per_cell, 2].
-        '''
-        anchor_hw = []
-        for s in self.anchor_areas:
-            for ar in self.aspect_ratios:  # w/h = ar
-                h = sqrt(s/ar)
-                w = ar * h
-                for sr in self.scale_ratios:  # scale
-                    anchor_h = h*sr
-                    anchor_w = w*sr
-                    anchor_hw.append([anchor_h, anchor_w])
-        num_fms = len(self.anchor_areas)
-        return torch.Tensor(anchor_hw).view(num_fms, -1, 2)
-
-    def _get_anchor_boxes(self, input_size):
-        '''Compute anchor boxes for each feature map.
-
-        Args:
-          input_size: (tensor) model input size of (h, w).
-
-        Returns:
-          boxes: (list) anchor boxes for each feature map. Each of size [#anchors,4],
-                        where #anchors = fmw * fmh * #anchors_per_cell
-        '''
-        num_fms = len(self.anchor_areas)
-        fm_sizes = [(input_size/pow(2.,i+3)).ceil() for i in range(num_fms)]  # p3 -> p7 feature map sizes
-
-        boxes = []
-        for i in range(num_fms):
-            fm_size = fm_sizes[i]
-            grid_size = input_size / fm_size
-            fm_h, fm_w = int(fm_size[0]), int(fm_size[1])
-            xy = meshgrid(fm_h,fm_w) + 0.5  # [fm_h*fm_w, 2]
-            xy = (xy*grid_size).view(fm_w,fm_h,1,2).expand(fm_w,fm_h,9,2)
-            hw = self.anchor_wh[i].view(1,1,9,2).expand(fm_w,fm_h,9,2)
-            box = torch.cat([xy,hw], 3)  # [x,y,w,h]
-            boxes.append(box.view(-1,4))
-        return torch.cat(boxes, 0)
-
-    def encode(self, boxes, labels, input_size):
-        '''Encode target bounding boxes and class labels.
-
-        We obey the Faster RCNN box coder:
-          tx = (x - anchor_x) / anchor_w
-          ty = (y - anchor_y) / anchor_h
-          tw = log(w / anchor_w)
-          th = log(h / anchor_h)
-
-        Args:
-          boxes: (tensor) bounding boxes of (xmin,ymin,xmax,ymax), sized [#obj, 4].
-          labels: (tensor) object class labels, sized [#obj,].
-          input_size: (int/tuple) model input size of (w,h).
-
-        Returns:
-          loc_targets: (tensor) encoded bounding boxes, sized [#anchors,4].
-          cls_targets: (tensor) encoded class labels, sized [#anchors,].
-        '''
-        input_size = torch.Tensor([input_size,input_size]) if isinstance(input_size, int) \
-                     else torch.Tensor(input_size)
-        anchor_boxes = self._get_anchor_boxes(input_size)
-        boxes = change_box_order(boxes, 'xyxy2xywh')
-
-        ious = box_iou(anchor_boxes, boxes, order='xywh')
-        max_ious, max_ids = ious.max(1)
-        boxes = boxes[max_ids]
-
-        loc_xy = (boxes[:,:2]-anchor_boxes[:,:2]) / anchor_boxes[:,2:]
-        loc_wh = torch.log(boxes[:,2:]/anchor_boxes[:,2:])
-        loc_targets = torch.cat([loc_xy,loc_wh], 1)
-        cls_targets = 1 + labels[max_ids]
-
-        cls_targets[max_ious<0.5] = 0
-        ignore = (max_ious>0.4) & (max_ious<0.5)  # ignore ious between [0.4,0.5]
-        cls_targets[ignore] = -1  # for now just mark ignored to -1
-        return loc_targets, cls_targets
-
-    def decode(self, loc_preds, cls_preds, input_size):
-        '''Decode outputs back to bouding box locations and class labels.
-
-        Args:
-          loc_preds: (tensor) predicted locations, sized [#anchors, 4].
-          cls_preds: (tensor) predicted class labels, sized [#anchors, #classes].
-          input_size: (int/tuple) model input size of (w,h).
-
-        Returns:
-          boxes: (tensor) decode box locations, sized [#obj,4].
-          labels: (tensor) class labels for each box, sized [#obj,].
-        '''
-        CLS_THRESH = 0.5
-        NMS_THRESH = 0.5
-
-        input_size = torch.Tensor([input_size,input_size]) if isinstance(input_size, int) \
-                     else torch.Tensor(input_size)
-        anchor_boxes = self._get_anchor_boxes(input_size)
-
-        loc_xy = loc_preds[:,:2]
-        loc_wh = loc_preds[:,2:]
-
-        xy = loc_xy * anchor_boxes[:,2:] + anchor_boxes[:,:2]
-        wh = loc_wh.exp() * anchor_boxes[:,2:]
-        boxes = torch.cat([xy-wh/2, xy+wh/2], 1)  # [#anchors,4]
-
-        score, labels = cls_preds.sigmoid().max(1)          # [#anchors,]
-        ids = score > CLS_THRESH
-        ids = ids.nonzero().squeeze()             # [#obj,]
-        keep = box_nms(boxes[ids], score[ids], threshold=NMS_THRESH)
-        return boxes[ids][keep], labels[ids][keep]
