+++ src/retinanet.py
@@ -3,8 +3,7 @@ import torch.nn as nn
 import torch.nn.functional as F
 from torch.autograd import Variable
 from torchvision import models
-
-from .utils import one_hot_embedding
+from math import sqrt
 
 
 class FPN(nn.Module):
@@ -53,7 +52,7 @@ class FPN(nn.Module):
         self.toplayer3 = nn.Conv2d(256, 256, kernel_size=3, stride=1, padding=1)
 
     def _upsample_add(self, x, y):
-        '''Upsample and add two feature maps.
+        """Upsample and add two feature maps.
 
         Args:
           x: (Variable) top feature map to be upsampled.
@@ -72,7 +71,7 @@ class FPN(nn.Module):
         upsampled feature map size: [N,_,16,16]
 
         So we choose bilinear upsample which supports arbitrary output sizes.
-        '''
+        """
         _, _, H, W = y.size()
         return F.upsample(x, size=(H, W), mode='bilinear') + y
 
@@ -103,12 +102,14 @@ class RetinaNet(nn.Module):
     """PyTorch RetinaNet model using ResNet(34, 50, 101 or 152 backbone).
 
     Paper: https://arxiv.org/pdf/1708.02002.pdf
-    Original PyTorch implementation: https://github.com/kuangliu/pytorch-retinanet
 
     Args:
         encoder_depth (int): Depth of ResNet backbone, must be in {34, 50, 101, 152}.
         pretrained_encoder (bool): If True weights pretrained on ImageNet will be used for ResNet. Defaults to True.
         num_classes (int): Number of classes for detection. Defaults to 20.
+
+    References:
+        https://github.com/kuangliu/pytorch-retinanet
     """
     num_anchors = 9
 
@@ -143,7 +144,7 @@ class RetinaNet(nn.Module):
         return nn.Sequential(*layers)
 
     def freeze_bn(self):
-        '''Freeze BatchNorm layers.'''
+        """Freeze BatchNorm layers."""
         for layer in self.modules():
             if isinstance(layer, nn.BatchNorm2d):
                 layer.eval()
@@ -155,7 +156,7 @@ class FocalLoss(nn.Module):
         self.num_classes = num_classes
 
     def focal_loss(self, x, y):
-        '''Focal loss.
+        """Focal loss.
 
         Args:
           x: (tensor) sized [N,D].
@@ -163,7 +164,7 @@ class FocalLoss(nn.Module):
 
         Return:
           (tensor) focal loss.
-        '''
+        """
         alpha = 0.25
         gamma = 2
 
@@ -178,7 +179,7 @@ class FocalLoss(nn.Module):
         return F.binary_cross_entropy_with_logits(x, t, w, size_average=False)
 
     def focal_loss_alt(self, x, y):
-        '''Focal loss alternative.
+        """Focal loss alternative.
 
         Args:
           x: (tensor) sized [N,D].
@@ -186,7 +187,7 @@ class FocalLoss(nn.Module):
 
         Return:
           (tensor) focal loss.
-        '''
+        """
         alpha = 0.25
 
         t = one_hot_embedding(y.data.cpu(), 1+self.num_classes)
@@ -201,7 +202,7 @@ class FocalLoss(nn.Module):
         return loss.sum()
 
     def forward(self, loc_preds, loc_targets, cls_preds, cls_targets):
-        '''Compute loss between (loc_preds, loc_targets) and (cls_preds, cls_targets).
+        """Compute loss between (loc_preds, loc_targets) and (cls_preds, cls_targets).
 
         Args:
           loc_preds: (tensor) predicted locations, sized [batch_size, #anchors, 4].
@@ -211,8 +212,7 @@ class FocalLoss(nn.Module):
 
         loss:
           (tensor) loss = SmoothL1Loss(loc_preds, loc_targets) + FocalLoss(cls_preds, cls_targets).
-        '''
-        batch_size, num_boxes = cls_targets.size()
+        """
         pos = cls_targets > 0  # [N,#anchors]
         num_pos = pos.data.long().sum()
 
@@ -237,15 +237,290 @@ class FocalLoss(nn.Module):
         return loss
 
 
-if __name__ == "__main__":
-    num_classes = 500
-    net = RetinaNet(encoder_depth=50,
-                    pretrained_encoder=True,
-                    num_classes=num_classes)
-    loc_preds, cls_preds = net(Variable(torch.randn(2, 3, 256, 256)))
-    print(loc_preds.size())
-    print(cls_preds.size())
-    loc_grads = Variable(torch.randn(loc_preds.size()))
-    cls_grads = Variable(torch.randn(cls_preds.size()))
-    loc_preds.backward(loc_grads)
-    cls_preds.backward(cls_grads)
+class DataEncoder(object):
+    def __init__(self):
+        self.anchor_areas = [32*32., 64*64., 128*128., 256*256., 512*512.]  # p3 -> p7
+        self.aspect_ratios = [1/2., 1/1., 2/1.]
+        self.scale_ratios = [1., pow(2,1/3.), pow(2,2/3.)]
+        self.anchor_wh = self._get_anchor_hw()
+
+    def _get_anchor_hw(self):
+        """Compute anchor height and width for each feature map.
+
+        Returns:
+          anchor_hw: (tensor) anchor hw, sized [#fm, #anchors_per_cell, 2].
+        """
+        anchor_hw = []
+        for s in self.anchor_areas:
+            for ar in self.aspect_ratios:  # w/h = ar
+                h = sqrt(s/ar)
+                w = ar * h
+                for sr in self.scale_ratios:  # scale
+                    anchor_h = h*sr
+                    anchor_w = w*sr
+                    anchor_hw.append([anchor_h, anchor_w])
+        num_fms = len(self.anchor_areas)
+        return torch.Tensor(anchor_hw).view(num_fms, -1, 2)
+
+    def _get_anchor_boxes(self, input_size):
+        """Compute anchor boxes for each feature map.
+
+        Args:
+          input_size: (tensor) model input size of (h, w).
+
+        Returns:
+          boxes: (list) anchor boxes for each feature map. Each of size [#anchors,4],
+                        where #anchors = fmw * fmh * #anchors_per_cell
+        """
+        num_fms = len(self.anchor_areas)
+        fm_sizes = [(input_size/pow(2.,i+3)).ceil() for i in range(num_fms)]  # p3 -> p7 feature map sizes
+
+        boxes = []
+        for i in range(num_fms):
+            fm_size = fm_sizes[i]
+            grid_size = input_size / fm_size
+            fm_h, fm_w = int(fm_size[0]), int(fm_size[1])
+            xy = meshgrid(fm_h,fm_w) + 0.5  # [fm_h*fm_w, 2]
+            xy = (xy*grid_size).view(fm_w,fm_h,1,2).expand(fm_w,fm_h,9,2)
+            hw = self.anchor_wh[i].view(1,1,9,2).expand(fm_w,fm_h,9,2)
+            box = torch.cat([xy,hw], 3)  # [x,y,w,h]
+            boxes.append(box.view(-1,4))
+        return torch.cat(boxes, 0)
+
+    def encode(self, boxes, labels, input_size):
+        """Encode target bounding boxes and class labels.
+
+        We obey the Faster RCNN box coder:
+          tx = (x - anchor_x) / anchor_w
+          ty = (y - anchor_y) / anchor_h
+          tw = log(w / anchor_w)
+          th = log(h / anchor_h)
+
+        Args:
+          boxes: (tensor) bounding boxes of (xmin,ymin,xmax,ymax), sized [#obj, 4].
+          labels: (tensor) object class labels, sized [#obj,].
+          input_size: (int/tuple) model input size of (h,w).
+
+        Returns:
+          loc_targets: (tensor) encoded bounding boxes, sized [#anchors,4].
+          cls_targets: (tensor) encoded class labels, sized [#anchors,].
+        """
+        input_size = torch.Tensor([input_size,input_size]) if isinstance(input_size, int) \
+                     else torch.Tensor(input_size)
+        anchor_boxes = self._get_anchor_boxes(input_size)
+        boxes = change_box_order(boxes, 'xyxy2xywh')
+
+        ious = box_iou(anchor_boxes, boxes, order='xywh')
+        max_ious, max_ids = ious.max(1)
+        boxes = boxes[max_ids]
+
+        loc_xy = (boxes[:,:2]-anchor_boxes[:,:2]) / anchor_boxes[:,2:]
+        loc_hw = torch.log(boxes[:,2:]/anchor_boxes[:,2:])
+        loc_targets = torch.cat([loc_xy,loc_hw], 1)
+        cls_targets = 1 + labels[max_ids]
+
+        cls_targets[max_ious<0.5] = 0
+        ignore = (max_ious>0.4) & (max_ious<0.5)  # ignore ious between [0.4,0.5]
+        cls_targets[ignore] = -1  # for now just mark ignored to -1
+        return loc_targets, cls_targets
+
+    def decode(self, loc_preds, cls_preds, input_size):
+        """Decode outputs back to bouding box locations and class labels.
+
+        Args:
+          loc_preds: (tensor) predicted locations, sized [#anchors, 4].
+          cls_preds: (tensor) predicted class labels, sized [#anchors, #classes].
+          input_size: (int/tuple) model input size of (h,w).
+
+        Returns:
+          boxes: (tensor) decode box locations, sized [#obj,4].
+          labels: (tensor) class labels for each box, sized [#obj,].
+        """
+        CLS_THRESH = 0.5
+        NMS_THRESH = 0.5
+
+        input_size = torch.Tensor([input_size,input_size]) if isinstance(input_size, int) \
+                     else torch.Tensor(input_size)
+        anchor_boxes = self._get_anchor_boxes(input_size)
+
+        loc_xy = loc_preds[:,:2]
+        loc_hw = loc_preds[:,2:]
+
+        xy = loc_xy * anchor_boxes[:,2:] + anchor_boxes[:,:2]
+        wh = loc_hw.exp() * anchor_boxes[:,2:]
+        boxes = torch.cat([xy-wh/2, xy+wh/2], 1)  # [#anchors,4]
+
+        score, labels = cls_preds.sigmoid().max(1)          # [#anchors,]
+        ids = score > CLS_THRESH
+        ids = ids.nonzero().squeeze()             # [#obj,]
+        keep = box_nms(boxes[ids], score[ids], threshold=NMS_THRESH)
+        return boxes[ids][keep], labels[ids][keep]
+
+
+def one_hot_embedding(labels, num_classes):
+    """Embedding labels to one-hot form.
+
+    Args:
+      labels: (LongTensor) class labels, sized [N,].
+      num_classes: (int) number of classes.
+
+    Returns:
+      (tensor) encoded labels, sized [N,#classes].
+
+    Reference:
+      https://github.com/kuangliu/pytorch-retinanet
+    """
+    y = torch.eye(num_classes)  # [D,D]
+    return y[labels]            # [N,D]
+
+
+def meshgrid(x, y, row_major=True):
+    """Return meshgrid in range x & y.
+
+    Args:
+      x: (int) first dim range.
+      y: (int) second dim range.
+      row_major: (bool) row major or column major.
+
+    Returns:
+      (tensor) meshgrid, sized [x*y,2]
+
+    Example:
+    >> meshgrid(3,2)
+    0  0
+    1  0
+    2  0
+    0  1
+    1  1
+    2  1
+    [torch.FloatTensor of size 6x2]
+
+    >> meshgrid(3,2,row_major=False)
+    0  0
+    0  1
+    0  2
+    1  0
+    1  1
+    1  2
+    [torch.FloatTensor of size 6x2]
+
+    Reference:
+      https://github.com/kuangliu/pytorch-retinanet
+    """
+    a = torch.arange(0,x)
+    b = torch.arange(0,y)
+    xx = a.repeat(y).view(-1,1)
+    yy = b.view(-1,1).repeat(1,x).view(-1,1)
+    return torch.cat([xx,yy],1) if row_major else torch.cat([yy,xx],1)
+
+
+def change_box_order(boxes, order):
+    """Change box order between (xmin,ymin,xmax,ymax) and (xcenter,ycenter,width,height).
+
+    Args:
+      boxes: (tensor) bounding boxes, sized [N,4].
+      order: (str) either 'xyxy2xywh' or 'xywh2xyxy'.
+
+    Returns:
+      (tensor) converted bounding boxes, sized [N,4].
+
+    Reference:
+      https://github.com/kuangliu/pytorch-retinanet
+    """
+    assert order in ['xyxy2xywh','xywh2xyxy']
+    a = boxes[:,:2]
+    b = boxes[:,2:]
+    if order == 'xyxy2xywh':
+        return torch.cat([(a+b)/2,b-a+1], 1)
+    return torch.cat([a-b/2,a+b/2], 1)
+
+
+def box_iou(box1, box2, order='xyxy'):
+    """Compute the intersection over union of two set of boxes.
+
+    The default box order is (xmin, ymin, xmax, ymax).
+
+    Args:
+      box1: (tensor) bounding boxes, sized [N,4].
+      box2: (tensor) bounding boxes, sized [M,4].
+      order: (str) box order, either 'xyxy' or 'xywh'.
+
+    Return:
+      (tensor) iou, sized [N,M].
+
+    Reference:
+      https://github.com/kuangliu/pytorch-retinanet
+      https://github.com/chainer/chainercv/blob/master/chainercv/utils/bbox/bbox_iou.py
+    """
+    if order == 'xywh':
+        box1 = change_box_order(box1, 'xywh2xyxy')
+        box2 = change_box_order(box2, 'xywh2xyxy')
+
+    N = box1.size(0)
+    M = box2.size(0)
+
+    lt = torch.max(box1[:,None,:2], box2[:,:2])  # [N,M,2]
+    rb = torch.min(box1[:,None,2:], box2[:,2:])  # [N,M,2]
+
+    wh = (rb-lt+1).clamp(min=0)      # [N,M,2]
+    inter = wh[:,:,0] * wh[:,:,1]  # [N,M]
+
+    area1 = (box1[:,2]-box1[:,0]+1) * (box1[:,3]-box1[:,1]+1)  # [N,]
+    area2 = (box2[:,2]-box2[:,0]+1) * (box2[:,3]-box2[:,1]+1)  # [M,]
+    iou = inter / (area1[:,None] + area2 - inter)
+    return iou
+
+
+def box_nms(bboxes, scores, threshold=0.5, mode='union'):
+    """Non maximum suppression.
+    source: https://github.com/kuangliu/pytorch-retinanet
+
+    Args:
+      bboxes: (tensor) bounding boxes, sized [N,4].
+      scores: (tensor) bbox scores, sized [N,].
+      threshold: (float) overlap threshold.
+      mode: (str) 'union' or 'min'.
+
+    Returns:
+      keep: (tensor) selected indices.
+
+    Reference:
+      https://github.com/rbgirshick/py-faster-rcnn/blob/master/lib/nms/py_cpu_nms.py
+    """
+    x1 = bboxes[:,0]
+    y1 = bboxes[:,1]
+    x2 = bboxes[:,2]
+    y2 = bboxes[:,3]
+
+    areas = (x2-x1+1) * (y2-y1+1)
+    _, order = scores.sort(0, descending=True)
+
+    keep = []
+    while order.numel() > 0:
+        i = order[0]
+        keep.append(i)
+
+        if order.numel() == 1:
+            break
+
+        xx1 = x1[order[1:]].clamp(min=x1[i])
+        yy1 = y1[order[1:]].clamp(min=y1[i])
+        xx2 = x2[order[1:]].clamp(max=x2[i])
+        yy2 = y2[order[1:]].clamp(max=y2[i])
+
+        w = (xx2-xx1+1).clamp(min=0)
+        h = (yy2-yy1+1).clamp(min=0)
+        inter = w*h
+
+        if mode == 'union':
+            ovr = inter / (areas[i] + areas[order[1:]] - inter)
+        elif mode == 'min':
+            ovr = inter / areas[order[1:]].clamp(max=areas[i])
+        else:
+            raise TypeError('Unknown nms mode: %s.' % mode)
+
+        ids = (ovr<=threshold).nonzero().squeeze()
+        if ids.numel() == 0:
+            break
+        order = order[ids+1]
+    return torch.LongTensor(keep)
