+++ src/utils.py
@@ -426,175 +426,3 @@ def make_apply_transformer_stream(func, output_name='output', apply_on=None):
 
     return StaticApplyTransformerStream()
 
-
-def get_seed():
-    seed = int(time.time()) + int(os.getpid())
-    return seed
-
-
-def reseed(augmenter_sequence, deterministic=True):
-    for aug in augmenter_sequence:
-        aug.random_state = ia.new_random_state(get_seed())
-        if deterministic:
-            aug.deterministic = True
-    return augmenter_sequence
-
-
-def one_hot_embedding(labels, num_classes):
-    '''Embedding labels to one-hot form.
-    source: https://github.com/kuangliu/pytorch-retinanet
-
-    Args:
-      labels: (LongTensor) class labels, sized [N,].
-      num_classes: (int) number of classes.
-
-    Returns:
-      (tensor) encoded labels, sized [N,#classes].
-    '''
-    y = torch.eye(num_classes)  # [D,D]
-    return y[labels]            # [N,D]
-
-
-def meshgrid(x, y, row_major=True):
-    '''Return meshgrid in range x & y.
-    source: https://github.com/kuangliu/pytorch-retinanet
-
-    Args:
-      x: (int) first dim range.
-      y: (int) second dim range.
-      row_major: (bool) row major or column major.
-
-    Returns:
-      (tensor) meshgrid, sized [x*y,2]
-
-    Example:
-    >> meshgrid(3,2)
-    0  0
-    1  0
-    2  0
-    0  1
-    1  1
-    2  1
-    [torch.FloatTensor of size 6x2]
-
-    >> meshgrid(3,2,row_major=False)
-    0  0
-    0  1
-    0  2
-    1  0
-    1  1
-    1  2
-    [torch.FloatTensor of size 6x2]
-    '''
-    a = torch.arange(0,x)
-    b = torch.arange(0,y)
-    xx = a.repeat(y).view(-1,1)
-    yy = b.view(-1,1).repeat(1,x).view(-1,1)
-    return torch.cat([xx,yy],1) if row_major else torch.cat([yy,xx],1)
-
-def change_box_order(boxes, order):
-    '''Change box order between (xmin,ymin,xmax,ymax) and (xcenter,ycenter,width,height).
-    source: https://github.com/kuangliu/pytorch-retinanet
-
-    Args:
-      boxes: (tensor) bounding boxes, sized [N,4].
-      order: (str) either 'xyxy2xywh' or 'xywh2xyxy'.
-
-    Returns:
-      (tensor) converted bounding boxes, sized [N,4].
-    '''
-    assert order in ['xyxy2xywh','xywh2xyxy']
-    a = boxes[:,:2]
-    b = boxes[:,2:]
-    if order == 'xyxy2xywh':
-        return torch.cat([(a+b)/2,b-a+1], 1)
-    return torch.cat([a-b/2,a+b/2], 1)
-
-def box_iou(box1, box2, order='xyxy'):
-    '''Compute the intersection over union of two set of boxes.
-    source: https://github.com/kuangliu/pytorch-retinanet
-
-    The default box order is (xmin, ymin, xmax, ymax).
-
-    Args:
-      box1: (tensor) bounding boxes, sized [N,4].
-      box2: (tensor) bounding boxes, sized [M,4].
-      order: (str) box order, either 'xyxy' or 'xywh'.
-
-    Return:
-      (tensor) iou, sized [N,M].
-
-    Reference:
-      https://github.com/chainer/chainercv/blob/master/chainercv/utils/bbox/bbox_iou.py
-    '''
-    if order == 'xywh':
-        box1 = change_box_order(box1, 'xywh2xyxy')
-        box2 = change_box_order(box2, 'xywh2xyxy')
-
-    N = box1.size(0)
-    M = box2.size(0)
-
-    lt = torch.max(box1[:,None,:2], box2[:,:2])  # [N,M,2]
-    rb = torch.min(box1[:,None,2:], box2[:,2:])  # [N,M,2]
-
-    wh = (rb-lt+1).clamp(min=0)      # [N,M,2]
-    inter = wh[:,:,0] * wh[:,:,1]  # [N,M]
-
-    area1 = (box1[:,2]-box1[:,0]+1) * (box1[:,3]-box1[:,1]+1)  # [N,]
-    area2 = (box2[:,2]-box2[:,0]+1) * (box2[:,3]-box2[:,1]+1)  # [M,]
-    iou = inter / (area1[:,None] + area2 - inter)
-    return iou
-
-def box_nms(bboxes, scores, threshold=0.5, mode='union'):
-    '''Non maximum suppression.
-    source: https://github.com/kuangliu/pytorch-retinanet
-
-    Args:
-      bboxes: (tensor) bounding boxes, sized [N,4].
-      scores: (tensor) bbox scores, sized [N,].
-      threshold: (float) overlap threshold.
-      mode: (str) 'union' or 'min'.
-
-    Returns:
-      keep: (tensor) selected indices.
-
-    Reference:
-      https://github.com/rbgirshick/py-faster-rcnn/blob/master/lib/nms/py_cpu_nms.py
-    '''
-    x1 = bboxes[:,0]
-    y1 = bboxes[:,1]
-    x2 = bboxes[:,2]
-    y2 = bboxes[:,3]
-
-    areas = (x2-x1+1) * (y2-y1+1)
-    _, order = scores.sort(0, descending=True)
-
-    keep = []
-    while order.numel() > 0:
-        i = order[0]
-        keep.append(i)
-
-        if order.numel() == 1:
-            break
-
-        xx1 = x1[order[1:]].clamp(min=x1[i])
-        yy1 = y1[order[1:]].clamp(min=y1[i])
-        xx2 = x2[order[1:]].clamp(max=x2[i])
-        yy2 = y2[order[1:]].clamp(max=y2[i])
-
-        w = (xx2-xx1+1).clamp(min=0)
-        h = (yy2-yy1+1).clamp(min=0)
-        inter = w*h
-
-        if mode == 'union':
-            ovr = inter / (areas[i] + areas[order[1:]] - inter)
-        elif mode == 'min':
-            ovr = inter / areas[order[1:]].clamp(max=areas[i])
-        else:
-            raise TypeError('Unknown nms mode: %s.' % mode)
-
-        ids = (ovr<=threshold).nonzero().squeeze()
-        if ids.numel() == 0:
-            break
-        order = order[ids+1]
-    return torch.LongTensor(keep)
